🚚MOVE : enum 파일 이동 및 댓글 도메인 위치 변경
