--- aptner/src/main/java/com/fastcampus/aptner/post/common/domain/Comment.java
@@ -1,60 +1,0 @@
-package com.fastcampus.aptner.post.common.domain;
-
-import com.fastcampus.aptner.member.domain.Member;
-import com.fastcampus.aptner.post.announcement.domain.Announcement;
-import com.fastcampus.aptner.post.communication.domain.Communication;
-import com.fastcampus.aptner.post.complaint.domain.Complaint;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@ToString
-@Table(name = "comment")
-@Entity
-public class Comment {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long commentId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member memberId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "communication_id")
-    private Communication communicationId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "complaint_id")
-    private Complaint complaintId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "announcement_id")
-    private Announcement announcementId;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "board_type", nullable = false)
-    private BoardType boardType;
-
-    @Column(name = "contents", nullable = false)
-    private String contents;
-
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-
-    @Column(name = "modified_at", length = 50, nullable = false)
-    private LocalDateTime modifiedAt;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "status", nullable = false)
-    private PostStatus status;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "parent_id")
-    private Comment parentId;
-}
