+++ aptner/src/main/java/com/fastcampus/aptner/post/communication/domain/Communication.java
@@ -1,7 +1,7 @@
 package com.fastcampus.aptner.post.communication.domain;
 
 import com.fastcampus.aptner.member.domain.Member;
-import com.fastcampus.aptner.post.common.domain.PostStatus;
+import com.fastcampus.aptner.post.common.enumType.PostStatus;
 import jakarta.persistence.*;
 import lombok.*;
