+++ aptner/src/main/java/com/fastcampus/aptner/post/complaint/domain/Complaint.java
@@ -1,8 +1,9 @@
 package com.fastcampus.aptner.post.complaint.domain;
 
+import com.fastcampus.aptner.global.handler.common.BaseTimeEntity;
 import com.fastcampus.aptner.member.domain.Member;
-import com.fastcampus.aptner.post.common.domain.Comment;
-import com.fastcampus.aptner.post.common.domain.PostStatus;
+import com.fastcampus.aptner.post.opinion.domain.Comment;
+import com.fastcampus.aptner.post.common.enumType.PostStatus;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.*;
@@ -17,7 +18,7 @@ import java.util.List;
 @ToString
 @Table(name = "compaint")
 @Entity
-public class Complaint {
+public class Complaint extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
